backend/requirements.txt
fastapi==0.104.1
requests>=2.31.0
uvicorn==0.24.0.post1
pymongo==4.5.0
pydantic==2.10.4
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic-settings==2.5.2
schedule==1.2.1